# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Go Shop Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/products": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new product with the provided details including categories\n## Fields Description\n- **name** (required): Product name, 2-100 characters\n- **description** (optional): Product description, 5-500 characters if provided\n- **price** (required): Product price, must be greater than 0\n- **stock** (required): Initial stock quantity, must be greater than 0\n- **is_active** (optional): Whether product is active for sale. Defaults to true if not specified\n## Example Request\n```json\n{\n\"name\": \"Premium Laptop\",\n\"description\": \"High-performance laptop with 16GB RAM and 512GB SSD\",\n\"price\": 1299.99,\n\"stock\": 50,\n\"is_active\": true,\n}\n```",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create a new product",
                "parameters": [
                    {
                        "description": "Product creation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_internal_dto.ProductCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-go-shop-backend_internal_dto_ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data or validation failed",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (e.g., invalid token)",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    },
                    "403": {
                        "description": "Access denied (e.g., no access rights)",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    }
                }
            }
        },
        "/products/": {
            "get": {
                "description": "Get a paginated list of products with optional filtering by category and sorting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "List products with filtering and pagination",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 50,
                        "description": "Maximum number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "name",
                            "created_at",
                            "price"
                        ],
                        "type": "string",
                        "description": "Field to sort by",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Sort in descending order",
                        "name": "order_desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-go-shop-backend_pkg_response_PaginatedResponse-array_go-shop-backend_internal_dto_ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    }
                }
            }
        },
        "/products/search": {
            "get": {
                "description": "Search products with full-text search capability on name and description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Search products",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query for full-text search on name and description fields",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 50,
                        "description": "Maximum number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-go-shop-backend_pkg_response_PaginatedResponse-array_go-shop-backend_internal_dto_ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Get detailed information about a specific product by its UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get product by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-go-shop-backend_internal_dto_ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partially update a product's information. Only provided fields will be updated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update an existing product",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Product update data (partial update)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_internal_dto.ProductUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-go-shop-backend_internal_dto_ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data or validation failed",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (e.g., invalid token)",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    },
                    "403": {
                        "description": "Access denied (e.g., no access rights)",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/go-shop-backend_pkg_response.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "go-shop-backend_internal_dto.ProductCategoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "go-shop-backend_internal_dto.ProductCreateRequest": {
            "type": "object",
            "required": [
                "name",
                "price",
                "stock"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 4096,
                    "minLength": 5
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "price": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "go-shop-backend_internal_dto.ProductResponse": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "integer"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-shop-backend_internal_dto.ProductCategoryResponse"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-shop-backend_internal_dto.UploadResponse"
                    }
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "reserved": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "go-shop-backend_internal_dto.ProductUpdateRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 4096,
                    "minLength": 5
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "price": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "go-shop-backend_internal_dto.UploadResponse": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "is_main": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "go-shop-backend_pkg_response.PaginatedResponse-array_go-shop-backend_internal_dto_ProductResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-shop-backend_internal_dto.ProductResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "go-shop-backend_pkg_response.Response-any": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                },
                "result": {}
            }
        },
...
        "go-shop-backend_pkg_response.Response-go-shop-backend_internal_dto_ProductResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/go-shop-backend_internal_dto.ProductResponse"
                }
            }
        },
...
        "go-shop-backend_pkg_response.Response-go-shop-backend_pkg_response_PaginatedResponse-array_go-shop-backend_internal_dto_ProductResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/go-shop-backend_pkg_response.PaginatedResponse-array_go-shop-backend_internal_dto_ProductResponse"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}